+++ src/main/java/com/getindata/kafka/connect/iceberg/sink/IcebergSinkConfiguration.java
@@ -26,18 +26,35 @@ public class IcebergSinkConfiguration {
     public static final String CATALOG_IMPL = ICEBERG_PREFIX + "catalog-impl";
     public static final String CATALOG_TYPE = ICEBERG_PREFIX + "type";
     private static final ConfigDef CONFIG_DEF = new ConfigDef()
-            .define(UPSERT, BOOLEAN, true, MEDIUM, "")
-            .define(UPSERT_KEEP_DELETES, BOOLEAN, true, MEDIUM, "")
-            .define(UPSERT_DEDUP_COLUMN, STRING, "__source_ts_ms", LOW, "")
-            .define(UPSERT_OP_COLUMN, STRING, "__op", LOW, "")
-            .define(ALLOW_FIELD_ADDITION, BOOLEAN, true, LOW, "")
-            .define(TABLE_AUTO_CREATE, BOOLEAN, false, MEDIUM, "")
-            .define(TABLE_NAMESPACE, STRING, "default", MEDIUM, "")
-            .define(TABLE_PREFIX, STRING, "", MEDIUM, "")
-            .define(TABLE_WRITE_FORMAT, STRING, "parquet", LOW, "")
-            .define(CATALOG_NAME, STRING, "default", MEDIUM, "")
-            .define(CATALOG_IMPL, STRING, null, MEDIUM, "")
-            .define(CATALOG_TYPE, STRING, null, MEDIUM, "")
+            .define(UPSERT, BOOLEAN, true, MEDIUM,
+                    "When true Iceberg rows will be updated based on table primary key. " +
+                            "When false all modification will be added as separate rows.")
+            .define(UPSERT_KEEP_DELETES, BOOLEAN, true, MEDIUM,
+                    "When true delete operation will leave a tombstone that will have only " +
+                            "a primary key and __deleted* flag set to true")
+            .define(UPSERT_DEDUP_COLUMN, STRING, "__source_ts_ms", LOW,
+                    "Column used to check which state is newer during upsert")
+            .define(UPSERT_OP_COLUMN, STRING, "__op", LOW,
+                    "Column used to check which state is newer during upsert when " +
+                            "upsert.dedup-column is not enough to resolve")
+            .define(ALLOW_FIELD_ADDITION, BOOLEAN, true, LOW,
+                    "When true sink will be adding new columns to Iceberg tables on schema changes")
+            .define(TABLE_AUTO_CREATE, BOOLEAN, false, MEDIUM,
+                    "When true sink will automatically create new Iceberg tables")
+            .define(TABLE_NAMESPACE, STRING, "default", MEDIUM,
+                    "Table namespace. In Glue it will be used as database name")
+            .define(TABLE_PREFIX, STRING, "", MEDIUM,
+                    "Prefix added to all table names")
+            .define(TABLE_WRITE_FORMAT, STRING, "parquet", LOW,
+                    "Format used for Iceberg tables")
+            .define(CATALOG_NAME, STRING, "default", MEDIUM,
+                    "Iceberg catalog name")
+            .define(CATALOG_IMPL, STRING, null, MEDIUM,
+                    "Iceberg catalog implementation (Only one of iceberg.catalog-impl and iceberg.type " +
+                            "can be set to non null value at the same time")
+            .define(CATALOG_TYPE, STRING, null, MEDIUM,
+                    "Iceberg catalog type (Only one of iceberg.catalog-impl and iceberg.type " +
+                            "can be set to non null value at the same time)")
             ;
 
     private final AbstractConfig parsedConfig;