+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+- Logger levels changes
+- Added documentation to sink configuration
+
 ## [0.1.2] - 2022-03-25
 
 ## [0.1.1] - 2022-03-25